# docs/swagger.yaml
basePath: /v1
definitions:
...
  internal.Cinema:
    properties:
      id:
        type: integer
      location:
        type: string
      name:
        type: string
      ower_id:
        type: integer
      version:
        type: integer
    type: object
  internal.Hall:
    properties:
      cinema_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      seat_arrangement:
        type: string
      seat_price:
        type: number
      version:
        type: integer
    type: object
...
  internal.Seat:
    properties:
      coordinates:
        type: string
      hall_id:
        type: integer
      id:
        type: integer
      version:
        type: integer
    type: object
...
  main.CreateCinemaResponse:
    properties:
      cinema:
        $ref: '#/definitions/internal.Cinema'
    type: object
  main.CreateHallResponse:
    properties:
      hall:
        $ref: '#/definitions/internal.Hall'
    type: object
...
  main.CreateSeatReponse:
    properties:
      seat:
        $ref: '#/definitions/internal.Seat'
    type: object
...
  main.GetCinemaResponse:
    properties:
      cinema:
        $ref: '#/definitions/internal.Cinema'
    type: object
  main.GetHallsResponse:
    properties:
      halls:
        items:
          $ref: '#/definitions/internal.Hall'
        type: array
    type: object
...
  main.ResponseError:
    properties:
      error:
        description: Error
        type: string
    type: object
  main.ResponseMessage:
    properties:
      message:
        description: Message
        type: string
    type: object
...
  main.UpdateHallResponse:
    properties:
      hall:
        $ref: '#/definitions/internal.Hall'
    type: object
...
  main.ViolationsMessage:
    properties:
      errors:
        additionalProperties:
          type: string
        description: errors
        type: object
    type: object
host: https://localhost:8080
info:
  contact:
    email: ahamerdev@gmail.com
    name: Ahmed Amer
  description: a simple movie reservation system api for educational purposes
  title: Movie Reservation System API
  version: "1.0"
paths:
...
  /cinemas:
    get:
      consumes:
      - application/json
      description: gets a list of cinemas by search parameters
      parameters:
      - description: name
        in: query
        name: name
        type: string
      - description: location
        in: query
        name: location
        type: string
      - description: page number
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: page_size
        type: integer
      - description: sort params are (name, location) prefix with - to sort descending
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.CreateCinemaResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Get a list of cinemas
      tags:
      - cinemas
    post:
      consumes:
      - application/json
      description: creates a cinema
      parameters:
      - description: name
        in: body
        name: name
        required: true
        schema:
          type: string
      - description: location
        in: body
        name: location
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.CreateCinemaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ViolationsMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Creates a cinema
      tags:
      - cinemas
  /cinemas/{id}:
    get:
      consumes:
      - application/json
      description: gets a cinema by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.GetCinemaResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Gets a cinema
      tags:
      - cinemas
    put:
      consumes:
      - application/json
      description: deletes a cinema
      parameters:
      - description: name
        in: body
        name: name
        schema:
          type: string
      - description: location
        in: body
        name: location
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Deletes a cinema
      tags:
      - cinemas
  /cinemas/{id}/halls:
    get:
      consumes:
      - application/json
      description: gets a list of halls for a given cinema
      parameters:
      - description: cinema id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.GetHallsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Gets a list of halls
      tags:
      - halls
    post:
      consumes:
      - application/json
      description: creates a hall for a given cinema
      parameters:
      - description: cinema id
        in: path
        name: id
        required: true
        type: integer
      - description: name
        in: body
        name: name
        schema:
          type: string
      - description: seat arrangement
        in: body
        name: seat_arrangement
        schema:
          type: string
      - description: seat price
        in: body
        name: seat_price
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.CreateHallResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ViolationsMessage'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Creates a hall
      tags:
      - halls
  /halls/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a hall by id
      parameters:
      - description: hall id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ViolationsMessage'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Deletes a hall
      tags:
      - halls
    put:
      consumes:
      - application/json
      description: Updates a hall by id
      parameters:
      - description: hall id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.UpdateHallResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ViolationsMessage'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Updates a hall
      tags:
      - halls
  /halls/{id}/seats:
    get:
      responses: {}
    post:
      consumes:
      - application/json
      description: Creates a seat for a given hall
      parameters:
      - description: hall id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.CreateSeatReponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ViolationsMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Creates a seat
      tags:
      - seats
...
  /seats/{id}:
    delete:
      consumes:
      - application/json
      description: deletes a seat by id
      parameters:
      - description: seat id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ViolationsMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Deletes a seat
      tags:
      - seats
    put:
      consumes:
      - application/json
      description: updates a seat by id
      parameters:
      - description: seat id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.CreateSeatReponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ViolationsMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Updates a seat
      tags:
      - seats
...
swagger: "2.0"